Moved all feature-related folders to the folder "feature"
