+++ lib/feature/data/datasources/i_data_source.dart
@@ -1,12 +1,10 @@
-import 'package:todo_list_yandex_school_2024/data/models/task_model.dart';
+import 'package:todo_list_yandex_school_2024/feature/data/models/task_model.dart';
 
-abstract class ITaskRepository {
+abstract interface class IDataSource {
   Future<List<TaskModel>> getAllTasks(bool firstLaunch);
-  Future<void> updateTasks(List<TaskModel> task);
+  Future<List<TaskModel>> updateTasks(List<TaskModel> task);
   Future<TaskModel> getTask(String taskId);
   Future<TaskModel> addTask(TaskModel task);
   Future<TaskModel> changeTask(TaskModel task);
   Future<TaskModel> deleteTask(String taskId);
-
-  List<TaskModel> get listOfTasks;
 }
