+++ lib/feature/domain/i_task_repository.dart
@@ -1,10 +1,12 @@
-import 'package:todo_list_yandex_school_2024/data/models/task_model.dart';
+import 'package:todo_list_yandex_school_2024/feature/data/models/task_model.dart';
 
-abstract interface class IDataSource {
+abstract class ITaskRepository {
   Future<List<TaskModel>> getAllTasks(bool firstLaunch);
-  Future<List<TaskModel>> updateTasks(List<TaskModel> task);
+  Future<void> updateTasks(List<TaskModel> task);
   Future<TaskModel> getTask(String taskId);
   Future<TaskModel> addTask(TaskModel task);
   Future<TaskModel> changeTask(TaskModel task);
   Future<TaskModel> deleteTask(String taskId);
+
+  List<TaskModel> get listOfTasks;
 }
