+++ lib/feature/domain/todo_list_bloc/task_list_events.dart
@@ -1,4 +1,4 @@
-import 'package:todo_list_yandex_school_2024/data/models/task_model.dart';
+import 'package:todo_list_yandex_school_2024/feature/data/models/task_model.dart';
 
 sealed class TaskListEvent {}
