+++ lib/feature/presentation/main_screen/presentation/pages/main_screen.dart
@@ -1,13 +1,13 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:todo_list_yandex_school_2024/core/logger.dart';
-import 'package:todo_list_yandex_school_2024/data/datasources/local_data_source.dart';
-import 'package:todo_list_yandex_school_2024/data/models/task_model.dart';
-import 'package:todo_list_yandex_school_2024/domain/todo_list_bloc/task_list_bloc.dart';
-import 'package:todo_list_yandex_school_2024/domain/todo_list_bloc/task_list_events.dart';
-import 'package:todo_list_yandex_school_2024/domain/todo_list_bloc/task_list_states.dart';
-import 'package:todo_list_yandex_school_2024/presentation/edit_task_screen/edit_task_screen.dart';
-import 'package:todo_list_yandex_school_2024/presentation/main_screen/presentation/widgets/checkbox_line.dart';
+import 'package:todo_list_yandex_school_2024/feature/data/datasources/local_data_source.dart';
+import 'package:todo_list_yandex_school_2024/feature/data/models/task_model.dart';
+import 'package:todo_list_yandex_school_2024/feature/domain/todo_list_bloc/task_list_bloc.dart';
+import 'package:todo_list_yandex_school_2024/feature/domain/todo_list_bloc/task_list_events.dart';
+import 'package:todo_list_yandex_school_2024/feature/domain/todo_list_bloc/task_list_states.dart';
+import 'package:todo_list_yandex_school_2024/feature/presentation/edit_task_screen/edit_task_screen.dart';
+import 'package:todo_list_yandex_school_2024/feature/presentation/main_screen/presentation/widgets/checkbox_line.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:todo_list_yandex_school_2024/service_locator.dart';
