+++ lib/feature/presentation/main_screen/presentation/widgets/checkbox_line.dart
@@ -1,11 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:todo_list_yandex_school_2024/core/date_formatter.dart';
-import 'package:todo_list_yandex_school_2024/data/models/task_model.dart';
-import 'package:todo_list_yandex_school_2024/domain/todo_list_bloc/task_list_bloc.dart';
-import 'package:todo_list_yandex_school_2024/domain/todo_list_bloc/task_list_events.dart';
+import 'package:todo_list_yandex_school_2024/feature/data/models/task_model.dart';
+import 'package:todo_list_yandex_school_2024/feature/domain/todo_list_bloc/task_list_bloc.dart';
+import 'package:todo_list_yandex_school_2024/feature/domain/todo_list_bloc/task_list_events.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
-
 
 import '../../../edit_task_screen/edit_task_screen.dart';
 
@@ -66,7 +65,8 @@ class _CheckboxLineState extends State<CheckboxLine> {
                     ),
                     if (widget.task.deadline != null)
                       Text(
-                        formatDate(widget.task.deadline!, AppLocalizations.of(context)!.languageCode),
+                        formatDate(widget.task.deadline!,
+                            AppLocalizations.of(context)!.languageCode),
                         style: TextStyle(
                           color: Colors.grey,
                         ),
@@ -86,7 +86,9 @@ class _CheckboxLineState extends State<CheckboxLine> {
                               )));
                   if (newTask != null) {
                     setState(() {
-                      context.read<TaskListBloc>().add(UpdateTaskEvent(newTask));
+                      context
+                          .read<TaskListBloc>()
+                          .add(UpdateTaskEvent(newTask));
                     });
                   }
                 },
