+++ lib/main.dart
@@ -4,14 +4,14 @@ import 'package:device_info_plus/device_info_plus.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:provider/provider.dart';
-import 'package:todo_list_yandex_school_2024/data/task_repository.dart';
-import 'package:todo_list_yandex_school_2024/domain/todo_list_bloc/task_list_bloc.dart';
-import 'package:todo_list_yandex_school_2024/domain/todo_list_bloc/task_list_events.dart';
+import 'package:todo_list_yandex_school_2024/feature/data/task_repository.dart';
+import 'package:todo_list_yandex_school_2024/feature/domain/todo_list_bloc/task_list_bloc.dart';
+import 'package:todo_list_yandex_school_2024/feature/domain/todo_list_bloc/task_list_events.dart';
 import 'package:todo_list_yandex_school_2024/l10n/l10n.dart';
 import "package:todo_list_yandex_school_2024/service_locator.dart";
 import 'package:flutter/material.dart';
 import 'package:todo_list_yandex_school_2024/core/logger.dart';
-import 'package:todo_list_yandex_school_2024/presentation/main_screen/presentation/pages/main_screen.dart';
+import 'package:todo_list_yandex_school_2024/feature/presentation/main_screen/presentation/pages/main_screen.dart';
 import 'package:uuid/uuid.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
