+++ lib/service_locator.dart
@@ -1,9 +1,9 @@
 import 'package:device_info_plus/device_info_plus.dart';
 import 'package:dio/dio.dart';
 import 'package:get_it/get_it.dart';
-import 'package:todo_list_yandex_school_2024/data/datasources/local_data_source.dart';
-import 'package:todo_list_yandex_school_2024/data/datasources/remote_data_source.dart';
-import 'package:todo_list_yandex_school_2024/data/task_repository.dart';
+import 'package:todo_list_yandex_school_2024/feature/data/datasources/local_data_source.dart';
+import 'package:todo_list_yandex_school_2024/feature/data/datasources/remote_data_source.dart';
+import 'package:todo_list_yandex_school_2024/feature/data/task_repository.dart';
 import 'package:uuid/uuid.dart';
 
 final getIt = GetIt.instance;
@@ -13,7 +13,6 @@ void setupLocator() {
   getIt.registerLazySingleton<LocalDataSource>(() => LocalDataSource());
   getIt.registerLazySingleton<RemoteDataSource>(
       () => RemoteDataSource(getIt<Dio>()));
-      
 
   // Register repository
   getIt.registerLazySingleton<TaskRepository>(() =>
