+++ test/remote_data_source_test.dart
@@ -1,7 +1,7 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
-import 'package:todo_list_yandex_school_2024/data/datasources/remote_data_source.dart';
-import 'package:todo_list_yandex_school_2024/data/models/task_model.dart';
+import 'package:todo_list_yandex_school_2024/feature/data/datasources/remote_data_source.dart';
+import 'package:todo_list_yandex_school_2024/feature/data/models/task_model.dart';
 import 'package:dio/dio.dart';
 import 'mocks.mocks.dart';
